+++ example/app.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from "react";
 import ReactDOM from "react-dom";
 
-import { Icon, Info, Success, Tag, Warning } from "../src";
+import { Icon, Info, Success, Tag, Thumbnail, Warning } from "../src";
 import paths from "../src/icons.json";
 
 const Heading = ({ children }) => <h1>{"<"}{children}{">"}</h1>;
@@ -29,6 +29,15 @@ const App = () => (
     <Tag>Blue</Tag>
     <Tag color="gray">Gray</Tag>
     <Tag color="red">Red</Tag>
+
+    <Heading>Thumbnail image=string</Heading>
+    {[1, 2, 3, 4, 5].map(ident => (
+      <Thumbnail
+        key={ident}
+        image={`https://robohash.org/${ident}`}
+        title={`Robot ${ident}`}
+      />
+    ))}
 
     <Heading>Warning</Heading>
     <Warning>This is a warning.</Warning>