+++ src/index.js
@@ -2,6 +2,7 @@ import Icon from "./components/Icon";
 import Info from "./components/Info";
 import Success from "./components/Success";
 import Tag from "./components/Tag";
+import Thumbnail from "./components/Thumbnail";
 import Warning from "./components/Warning";
 
 export {
@@ -9,5 +10,6 @@ export {
   Info,
   Success,
   Tag,
+  Thumbnail,
   Warning
 };