+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - The Icon component, as well as the dev libraries needed to support dynamic import in webpack.
 - The Tag component.
+- The Thumbnail component.
 
 ## [0.0.3] - 2018-06-15
 ### Changed